+++ pysharedutils/__init__.py
@@ -31,6 +31,7 @@ __all__ = [
     'snake_case_dict_keys',
     'camel_case_dict_keys',
     'get_dict_properties',
+    'map_dict_keys',
     # Encodings
     'force_bytes',
     'force_str',