+++ pysharedutils/dicts.py
@@ -15,6 +15,7 @@ __all__ = [
     'snake_case_dict_keys',
     'camel_case_dict_keys',
     'get_dict_properties',
+    'map_dict_keys',
 ]
 
 
@@ -254,3 +255,40 @@ def get_dict_properties(obj, strict, *args):
         output[key] = value
 
     return output
+
+
+def map_dict_keys(obj, map_obj):
+    """
+    :param obj: A python dict object who's keys has to be mapped..
+
+    :param map_obj: A map dict specifying the key and the new key..
+
+    Example:
+        >>> obj = {'first_name': 'Foo', 'last_name': 'Bar'}
+        >>> map_obj = {'first_name': 'given_name'}
+        >>> map_dict_keys(obj, map_obj)
+        {'given_name': 'Foo', 'last_name': 'Bar'}
+    """
+    for key, value in six.iteritems(map_obj):
+        map_key = value
+        if '.' in key:
+            coppied_obj = copy.copy(obj)
+            sources = key.split('.')
+            sources_len = len(sources)
+            for index, source in enumerate(sources):
+                coppied_obj = coppied_obj.get(source)
+                # Call the recursive `map_dict_keys` when you hit the
+                # second last element in the source.
+                if coppied_obj and index == (sources_len - 2):
+                    # Create the map obj for the leaf obj
+                    leaf_map_obj = {
+                        sources[index + 1]: map_key
+                    }
+                    map_dict_keys(
+                        coppied_obj,
+                        leaf_map_obj
+                    )
+        else:
+            if key in obj:
+                obj[map_key] = obj.pop(key)
+    return obj